pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["RonaldoSiqueira <ronaldo.edicassio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<3.14.0"
fastapi = {extras = ["standard"], version = "^0.117.1"}
sqlalchemy = "^2.0.43"
pydantic-settings = "^2.11.0"
alembic = "^1.16.5"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.6"
httpx = "^0.28.1"
coverage = "^7.10.7"
harlequin = "^2.1.3"

[tool.ruff]
line-length = 150
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = [
    "--cov=fast_zero",
    "--cov-report=term-missing",
    "-s",
    "-x",
    "-vv",
    "-p", "no:warnings"
]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest'
post_test = 'pytest --cov-report=html --cov-report=term-missing'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"